requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.98.0
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
PyMuPDF==1.22.5
python-docx==1.1.0
spacy==3.7.2
nltk==3.8.1
httpx==0.28.1
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
# Performance dependencies
faiss-cpu==1.7.4
numpy==1.24.3
uvloop==0.17.0
httptools==0.6.1
diskcache==5.6.3
